add menu seting, plash loading dll

+++ app/src/main/java/com/garudatekno/jemaah/menu/ViewPanduanVideo.java
@@ -15,6 +15,7 @@ import android.os.Message;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.PopupMenu;
 import android.util.Log;
+import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Button;
@@ -54,6 +55,7 @@ import de.hdodenhof.circleimageview.CircleImageView;
 import me.anwarshahriar.calligrapher.Calligrapher;
 
 import static com.garudatekno.jemaah.app.AppConfig.URL_HOME;
+import static java.lang.Boolean.FALSE;
 
 public class ViewPanduanVideo extends AppCompatActivity implements View.OnClickListener {
 
@@ -93,6 +95,7 @@ public class ViewPanduanVideo extends AppCompatActivity implements View.OnClickL
         setContentView(R.layout.view_panduan_video);
         Calligrapher calligrapher=new Calligrapher(this);
         calligrapher.setFont(this,"fonts/helvetica.ttf",true);
+        session = new SessionManager(getApplicationContext());
         //HEADER
         TextView txt_emergency=(TextView) findViewById(R.id.txt_emergency);
         TextView txt_thowaf=(TextView) findViewById(R.id.txt_thowaf);
@@ -180,32 +183,56 @@ public class ViewPanduanVideo extends AppCompatActivity implements View.OnClickL
             }
         });
         final  ImageView img_setting=(ImageView) findViewById(R.id.img_setting);
+        final PopupMenu popup = new PopupMenu(this, img_setting);
+        popup.getMenuInflater().inflate(R.menu.popup_menu, popup.getMenu());
+        if (!session.isLoggedIn()) {
+            Menu popupMenu = popup.getMenu();
+            popupMenu.findItem(R.id.logout).setVisible(FALSE);
+        }
         img_setting.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                //Creating the instance of PopupMenu
-                PopupMenu popup = new PopupMenu(ViewPanduanVideo.this, img_setting);
-                //Inflating the Popup using xml file
-                popup.getMenuInflater().inflate(R.menu.popup_menu, popup.getMenu());
-                //registering popup with OnMenuItemClickListener
                 popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
                     public boolean onMenuItemClick(MenuItem item) {
                         int id = item.getItemId();
                         if(id == R.id.logout) {
                             logoutUser();
+                        }if(id == R.id.donasi) {
+                            Uri uriUrl = Uri.parse("https://kitabisa.com/gohaji");
+                            Intent launchBrowser = new Intent(Intent.ACTION_VIEW, uriUrl);
+                            startActivity(launchBrowser);
+                        }if(id == R.id.penilaian) {
+                            Intent i = new Intent(getApplicationContext(), PenilaianTravel.class);
+                            startActivity(i);
+                        }if(id == R.id.cek_visa) {
+                            Uri uriUrl = Uri.parse("https://eservices.haj.gov.sa/eservices3/pages/VisaInquiry/SearchVisa.xhtml?dswid=4963");
+                            Intent launchBrowser = new Intent(Intent.ACTION_VIEW, uriUrl);
+                            startActivity(launchBrowser);
+                        }if(id == R.id.share) {
+                            try {
+                                Intent i = new Intent(Intent.ACTION_SEND);
+                                i.setType("text/plain");
+                                i.putExtra(Intent.EXTRA_SUBJECT, "GoHajj");
+                                String sAux = "\nLet me recommend you this application\n\n";
+                                sAux = sAux + "https://play.google.com/store/apps/details?id=GoHajj.Soft \n\n";
+                                i.putExtra(Intent.EXTRA_TEXT, sAux);
+                                startActivity(Intent.createChooser(i, "choose one"));
+                            } catch(Exception e) {
+                                //e.toString();
+                            }
+                        }if(id == R.id.download_doa) {
+
                         }
                         return true;
                     }
                 });
-
                 popup.show();//showing popup menu
             }
         });
 
-        session = new SessionManager(getApplicationContext());
-        if (!session.isLoggedIn()) {
-            logoutUser();
-        }
+//        if (!session.isLoggedIn()) {
+//            logoutUser();
+//        }
 
         Intent intent = getIntent();
         id = intent.getStringExtra(AppConfig.EMP_ID);
@@ -227,6 +254,7 @@ public class ViewPanduanVideo extends AppCompatActivity implements View.OnClickL
             videoView.setVideoPath("/sdcard/android/data/com.garudatekno.jemaah/panduan/"+file);
         }
         videoView.requestFocus();
+        videoView.setScaleX((float) 1);
         MediaController mediaController = new
                 MediaController(this);
         mediaController.setAnchorView(videoView);
@@ -264,9 +292,10 @@ public class ViewPanduanVideo extends AppCompatActivity implements View.OnClickL
     }
 
     private void logoutUser() {
-        session.setLogin(false);
-
-        db.deleteUsers();
+        if (session.isLoggedIn()) {
+            session.setLogin(false);
+            db.deleteUsers();
+        }
 
         // Launching the login activity
         Intent intent = new Intent(ViewPanduanVideo.this, LoginActivity.class);
